requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
WTForms==3.2.1
selenium==4.34.2
requests==2.32.4
beautifulsoup4==4.13.4
python-dotenv==1.1.1
Bootstrap-Flask==2.5.0
SQLAlchemy==2.0.42
